update api utils.py
